+++ extensions/apple-reminders/swift/AppleReminders/Sources/Reminders.swift
@@ -42,6 +42,7 @@ enum RemindersError: Error {
   case accessDenied
   case noRemindersFound
   case noReminderFound
+  case noListFound
   case unableToSaveReminder
   case other
 }
@@ -165,7 +166,7 @@ struct Recurrence: Decodable {
   if let listId = newReminder.listId {
     let calendars = eventStore.calendars(for: .reminder)
     guard let calendar = (calendars.first { $0.calendarIdentifier == listId }) else {
-      throw RemindersError.noReminderFound
+      throw RemindersError.noListFound
     }
     reminder.calendar = calendar
   } else {
@@ -311,7 +312,7 @@ struct MoveToListPayload: Decodable {
 
   let calendars = eventStore.calendars(for: .reminder)
   guard let newCalendar = (calendars.first { $0.calendarIdentifier == payload.listId }) else {
-    throw RemindersError.noReminderFound
+    throw RemindersError.noListFound
   }
 
   item.calendar = newCalendar
@@ -497,6 +498,7 @@ struct SetLocationPayload: Decodable {
 
 struct UpdateReminderPayload: Decodable {
   let reminderId: String
+  let listId: String?
   let title: String?
   let notes: String?
   let dueDate: String?
@@ -512,6 +514,14 @@ struct UpdateReminderPayload: Decodable {
 
   guard let item = eventStore.calendarItem(withIdentifier: payload.reminderId) as? EKReminder else {
     throw RemindersError.noReminderFound
+  }
+
+  if let listId = payload.listId {
+    let calendars = eventStore.calendars(for: .reminder)
+    guard let newCalendar = (calendars.first { $0.calendarIdentifier == listId }) else {
+      throw RemindersError.noListFound
+    }
+    item.calendar = newCalendar
   }
 
   if let isCompleted = payload.isCompleted {